+++ jkit-sql/src/main/java/com/alianga/jkit/sql/SqlTokenType.java
@@ -241,6 +241,8 @@ public enum SqlTokenType {
     SIBLINGS,
     PIVOT,
     UNPIVOT,
+    MATCH,
+    AGAINST,
     MATCHED,
     SOURCE,
     TARGET,